+++ framework/json-domain/src/main/java/is/codion/framework/json/domain/EntityObjectMapper.java
@@ -18,6 +18,8 @@
  */
 package is.codion.framework.json.domain;
 
+import is.codion.common.db.operation.FunctionType;
+import is.codion.common.db.operation.ProcedureType;
 import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
@@ -36,11 +38,17 @@ import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.Serial;
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import static java.util.Objects.requireNonNull;
 
@@ -57,12 +65,14 @@ public final class EntityObjectMapper extends ObjectMapper {
 	public static final TypeReference<List<Entity.Key>> KEY_LIST_REFERENCE = new TypeReference<List<Entity.Key>>() {};
 	public static final TypeReference<List<Entity>> ENTITY_LIST_REFERENCE = new TypeReference<List<Entity>>() {};
 
+	private final Entities entities;
 	private final SimpleModule module;
 	private final EntitySerializer entitySerializer;
 	private final EntityDeserializer entityDeserializer;
 	private final ConditionSerializer conditionSerializer;
 	private final ConditionDeserializer conditionDeserializer;
-	private final Entities entities;
+	private final Map<ProcedureType<?, ?>, DefaultProcedureDefinition<?>> procedureDefinitions = new HashMap<>();
+	private final Map<FunctionType<?, ?, ?>, DefaultFunctionDefinition<?, ?>> functionDefinitions = new HashMap<>();
 
 	EntityObjectMapper(Entities entities) {
 		this.entities = requireNonNull(entities);
@@ -198,11 +208,254 @@ public final class EntityObjectMapper extends ObjectMapper {
 	}
 
 	/**
+	 * Define the argument type for a procedure
+	 * @param procedureType the procedure type
+	 * @param <T> the argument type
+	 * @return the {@link ProcedureArgumentTypeDefiner}
+	 * @throws IllegalStateException in case this procedure has already been defined
+	 */
+	public <T> ProcedureArgumentTypeDefiner<T> define(ProcedureType<?, T> procedureType) {
+		return new DefaultProcedureArgumentTypeDefiner<>(requireNonNull(procedureType));
+	}
+
+	/**
+	 * Define the return type for a function
+	 * @param functionType the function type
+	 * @param <T> the argument type
+	 * @param <R> the return type
+	 * @return the {@link FunctionArgumentTypeDefiner}
+	 * @throws IllegalStateException in case this function has already been defined
+	 */
+	public <T, R> FunctionReturnTypeDefiner<T, R> define(FunctionType<?, T, R> functionType) {
+		return new DefaultFunctionReturnTypeDefiner<>(requireNonNull(functionType));
+	}
+
+	/**
+	 * @param procedureType the procedure type
+	 * @param <T> the argument type
+	 * @return the procedure definition
+	 */
+	public <T> ProcedureDefinition<T> procedure(ProcedureType<?, T> procedureType) {
+		ProcedureDefinition<T> definition = (ProcedureDefinition<T>) procedureDefinitions.get(requireNonNull(procedureType));
+		if (definition == null) {
+			throw new IllegalArgumentException("Procedure not defined " + procedureType);
+		}
+
+		return definition;
+	}
+
+	/**
+	 * @param functionType the function type
+	 * @param <T> the argument type
+	 * @param <R> the return type
+	 * @return the function definition
+	 */
+	public <T, R> FunctionDefinition<T, R> function(FunctionType<?, T, R> functionType) {
+		FunctionDefinition<T, R> definition = (FunctionDefinition<T, R>) functionDefinitions.get(requireNonNull(functionType));
+		if (definition == null) {
+			throw new IllegalArgumentException("Function not defined " + functionType);
+		}
+
+		return definition;
+	}
+
+	/**
+	 * @param <T> the argument type
+	 */
+	public interface ProcedureDefinition<T> {
+
+		/**
+		 * @return the argument type
+		 */
+		Class<T> argumentType();
+	}
+
+	/**
+	 * @param <T> the argument type
+	 * @param <R> the return type
+	 */
+	public interface FunctionDefinition<T, R> {
+
+		/**
+		 * @return the argument type or an empty {@link Optional} in case none is defined
+		 */
+		Optional<Class<T>> argumentType();
+
+		/**
+		 * @return the return type
+		 */
+		Class<R> returnType();
+	}
+
+	/**
+	 * @param <T> the argument type
+	 */
+	public interface ProcedureArgumentTypeDefiner<T> {
+
+		/**
+		 * @param argumentType the argument type
+		 */
+		void argumentType(TypeReference<T> argumentType);
+
+		/**
+		 * @param argumentType the argument type
+		 */
+		void argumentType(Class<T> argumentType);
+	}
+
+	/**
+	 * @param <T> the argument type
+	 * @param <R> the return type
+	 */
+	public interface FunctionReturnTypeDefiner<T, R> {
+
+		/**
+		 * @param returnType the return type
+		 * @return the function definer
+		 */
+		FunctionArgumentTypeDefiner<T> returnType(TypeReference<R> returnType);
+
+		/**
+		 * @param returnType the return type
+		 * @return the function definer
+		 */
+		FunctionArgumentTypeDefiner<T> returnType(Class<R> returnType);
+	}
+
+	/**
+	 * @param <T> the argument type
+	 */
+	public interface FunctionArgumentTypeDefiner<T> {
+
+		/**
+		 * @param argumentType the argument type
+		 */
+		void argumentType(TypeReference<T> argumentType);
+
+		/**
+		 * @param argumentType the argument type
+		 */
+		void argumentType(Class<T> argumentType);
+	}
+
+	private final class DefaultProcedureArgumentTypeDefiner<T> implements ProcedureArgumentTypeDefiner<T> {
+
+		private final ProcedureType<?, T> procedureType;
+
+		private DefaultProcedureArgumentTypeDefiner(ProcedureType<?, T> procedureType) {
+			if (procedureDefinitions.containsKey(procedureType)) {
+				throw new IllegalStateException("Procedure already defined " + procedureType);
+			}
+			this.procedureType = procedureType;
+		}
+
+		@Override
+		public void argumentType(TypeReference<T> argumentType) {
+			argumentType(rawType(requireNonNull(argumentType)));
+		}
+
+		@Override
+		public void argumentType(Class<T> argumentType) {
+			procedureDefinitions.computeIfAbsent(procedureType, k -> new DefaultProcedureDefinition<>(requireNonNull(argumentType)));
+		}
+	}
+
+	private final class DefaultFunctionReturnTypeDefiner<T, R> implements FunctionReturnTypeDefiner<T, R> {
+
+		private final FunctionType<?, T, R> functionType;
+
+		private DefaultFunctionReturnTypeDefiner(FunctionType<?, T, R> functionType) {
+			if (functionDefinitions.containsKey(functionType)) {
+				throw new IllegalStateException("Function already defined " + functionType);
+			}
+			this.functionType = functionType;
+		}
+
+		@Override
+		public FunctionArgumentTypeDefiner<T> returnType(TypeReference<R> returnType) {
+			return returnType(rawType(requireNonNull(returnType)));
+		}
+
+		@Override
+		public FunctionArgumentTypeDefiner<T> returnType(Class<R> returnType) {
+			return new DefaultFunctionArgumentTypeDefiner<>(functionType, requireNonNull(returnType));
+		}
+	}
+
+	private final class DefaultFunctionArgumentTypeDefiner<T, R> implements FunctionArgumentTypeDefiner<T> {
+
+		private final FunctionType<?, T, R> functionType;
+
+		private DefaultFunctionArgumentTypeDefiner(FunctionType<?, T, R> functionType, Class<R> returnType) {
+			this.functionType = requireNonNull(functionType);
+			functionDefinitions.put(functionType, new DefaultFunctionDefinition<>(returnType, null));
+		}
+
+		@Override
+		public void argumentType(TypeReference<T> argumentType) {
+			argumentType(rawType(requireNonNull(argumentType)));
+		}
+
+		@Override
+		public void argumentType(Class<T> argumentType) {
+			functionDefinitions.compute(functionType, (k, definition) ->
+							new DefaultFunctionDefinition<>(definition.returnType, requireNonNull(argumentType)));
+		}
+	}
+
+	private static final class DefaultProcedureDefinition<T> implements ProcedureDefinition<T> {
+
+		private final Class<T> argumentType;
+
+		private DefaultProcedureDefinition(Class<T> argumentType) {
+			this.argumentType = argumentType;
+		}
+
+		@Override
+		public Class<T> argumentType() {
+			return argumentType;
+		}
+	}
+
+	private static final class DefaultFunctionDefinition<T, R> implements FunctionDefinition<T, R> {
+
+		private final Class<R> returnType;
+		private final @Nullable Class<T> argumentType;
+
+		private DefaultFunctionDefinition(Class<R> returnType, @Nullable Class<T> argumentType) {
+			this.returnType = returnType;
+			this.argumentType = argumentType;
+		}
+
+		@Override
+		public Class<R> returnType() {
+			return returnType;
+		}
+
+		@Override
+		public Optional<Class<T>> argumentType() {
+			return Optional.ofNullable(argumentType);
+		}
+	}
+
+	/**
 	 * A factory method for {@link EntityObjectMapper} instances.
 	 * @param entities the domain entities
 	 * @return a new {@link EntityObjectMapper} instance based on the given entities
 	 */
 	public static EntityObjectMapper entityObjectMapper(Entities entities) {
 		return new EntityObjectMapper(entities);
+	}
+
+	private static <T> Class<T> rawType(TypeReference<T> typeReference) {
+		Type type = typeReference.getType();
+		if (type instanceof ParameterizedType) {
+			return (Class<T>) ((ParameterizedType) type).getRawType();
+		}
+		else if (type instanceof Class<?>) {
+			return (Class<T>) type;
+		}
+
+		throw new IllegalArgumentException("Cannot extract raw type from: " + type);
 	}
 }